fix search/crypto

+++ models.py
@@ -30,7 +30,7 @@ class Company(Base, Image):
     instance_comex = models.BooleanField(_.instance_comex, default=False)
     matrix_skills = models.BooleanField(_.matrix_skills, default=False)
     
-    capital_socnomtotal = models.BigIntegerField(_.capital_socnomtotal, blank=True, null=True)
+    capital_socnomtotal = models.DecimalField(_.capital_socnomtotal, blank=True, null=True, decimal_places=2, max_digits=15)
     capitalisation = models.FloatField(_.capitalisation, default=0)
     turnover = models.BigIntegerField(_.turnover, blank=True, null=True)
     net_profit = models.BigIntegerField(_.net_profit, blank=True, null=True, help_text=_.net_profit_help)
